integer overflow in expression fix

+++ benchmark/micro/radixjoin.cpp
@@ -3,11 +3,12 @@
 #include "duckdb/main/appender.hpp"
 
 #include <random>
+#include <iostream>
 
 using namespace duckdb;
 using namespace std;
 
-#define RADIXJOIN_COUNT 256*1024
+#define RADIXJOIN_COUNT (size_t) 256*1024
 DUCKDB_BENCHMARK(RadixJoin, "[micro]")
     virtual void Load(DuckDBBenchmarkState *state) {
         // fixed seed random numbers
